Improve build/test scripts, CMake and HDF5
Consolidate and modernize build/test infrastructure and fix HDF5 integer typing.

- Replace and extend shell/batch helpers: add run.sh/run.bat; make setup, build and runtests accept a mode (debug|release) and operate on build-<mode> directories; remove old debug-specific scripts.
- Update setup/build/runtests scripts (sh and bat) to handle debug/release flows, vcpkg installation checks, and runtime output paths.
- Revamp CMakeLists: centralize include dirs, add per-target compile/link options for MSVC and non-MSVC (warnings and AddressSanitizer in Debug), remove older global compile option, and ensure tests inherit similar flags; minor formatting/layout fixes.
- HDF5/type fixes: switch HDF5 attribute/array types to use uint64_t (NATIVE_UINT64) and cast size_t uses to uint64_t when writing to HDF5 to avoid type mismatch.
- API and header cleanups: change several virtual getters to return size_t (instead of const size_t), normalize parameter names and spacing, reorder some member declarations, and update implementations to use size_t consistently.
- Other tweaks: remove redundant runtime checks in PaddingValues, update .gitignore to include .pdb and new build dir patterns, and minor code style/whitespace adjustments across multiple files.

These changes aim to provide clearer multi-configuration build/test workflows, safer HDF5 persistence for large indexes/sizes, and consistent types across the codebase.

+++ include/Optimizers/AdamBaseOptimizers/AdamBaseOptimizer.hpp
@@ -133,9 +133,9 @@ namespace Optimizers
                 size_t biasesSize = sizes.second;
                 std::string layerIdStr = std::to_string(layerId);
                 H5::Group layerGroup = momentBuffersGroup.createGroup(layerIdStr);
-                Utils::writeValueToHDF5<size_t>(layerGroup, "layerId", layerId);
-                Utils::writeValueToHDF5<size_t>(layerGroup, "weightsSize", weightsSize);
-                Utils::writeValueToHDF5<size_t>(layerGroup, "biasesSize", biasesSize);
+                Utils::writeValueToHDF5<uint64_t>(layerGroup, "layerId", static_cast<uint64_t>(layerId));
+                Utils::writeValueToHDF5<uint64_t>(layerGroup, "weightsSize", static_cast<uint64_t>(weightsSize));
+                Utils::writeValueToHDF5<uint64_t>(layerGroup, "biasesSize", static_cast<uint64_t>(biasesSize));
                 if (weightsSize > 0)
                 {
                     std::string weightKey = layerIdStr + "Weights";
